+++ lilac/src/main/java/com/hartwig/hmftools/lilac/candidates/NucleotideFiltering.java
@@ -14,10 +14,10 @@ import java.util.stream.Collectors;
 
 public class NucleotideFiltering
 {
-    private final int mMinNucleotideCount;
+    private final double mMinNucleotideCount;
     private final List<Integer> mAminoAcidBoundaries;
 
-    public NucleotideFiltering(int minNucleotideCount, final List<Integer> aminoAcidBoundaries)
+    public NucleotideFiltering(double minNucleotideCount, final List<Integer> aminoAcidBoundaries)
     {
         mMinNucleotideCount = minNucleotideCount;
         mAminoAcidBoundaries = aminoAcidBoundaries;
